Add support for features
Add feature options to:

- build
- check
- clippy
- doc
- run
- test

Options:

- `--features`
- `--all-features`
- `--no-default-features`

Also, change most of the argument key bindings to have a prefix based
on the name of the group that contains them. For example, `--bin`'s
key is changed from "-b" to "-cb" (it's in the "Compilation Options"
group, so it gets a "c" prefix). This is necessary to make room for
some of the feature arguments.

Top level arguments (eg. "Clippy Options" while in the "clippy"
transient), do not have prefixes, nor do the actions in the "Actions"
groups.

Refactor any group titles that are used more than once into variables.

Close #2