Add support for `--nocapture` in `cargo test`
As part of that support, add `cargo-transient--rearrange-args`, which
handles rearranging arguments in the arguments list if any of the
arguments need to follow a trailing `--` in the arguments passed to
`cargo` (eg. because they are actually arguments to the test binary,
rather than arguments to `cargo test` itself).

Also change all the main functions (`cargo-transient-build`, etc.) to
take their inputs as individual arguments rather than a single
argument which is a list of strings. That avoids
`cargo-transient--args` having to return its arguments wrappen in an
extra list.

Close #26