Replace use of `project-prefixed-buffer-name` with a custom function
An unintended consequence of 69b56ef was that
`project-prefixed-buffer-name` no longer worked as intended as the
value of `cargo-transient-compilation-buffer-name-function`, because
it relies on `default-directory` being set to the project root.

This commit replaces it as an option for
`cargo-transient-compilation-buffer-name-function` with a wrapped
version, specific to cargo-transient, which sets `default-directory`
before calling `project-prefixed-buffer-name` if it is run in the
context of a project.