Run cargo from the directory containing Cargo.toml
... if it can be located.  Otherwise use the current value of
`default-directory`, although it seems very unlikely that any cargo
commands would succeed in that case.

This improves on 69b56ef while retaining the same functionality.
Previously, cargo itself would look for Cargo.toml in parent
directories, but the compilation buffer was not able to open files
referenced in error messages because their paths were relative to the
Cargo.toml file, not the `default-directory` of the compilation
buffer.

Close #32